+++ iob_eth/software/simulation/src/iob_axistream_out_csrs.c
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: MIT
  *
- * Py2HWSW Version 0.81 has generated this code (https://github.com/IObundle/py2hwsw).
+ * Py2HWSW Version 0.81.0 has generated this code (https://github.com/IObundle/py2hwsw).
  */
 
 #include "iob_axistream_out_csrs.h"
@@ -58,7 +58,7 @@ uint32_t iob_axistream_out_csrs_get_fifo_level() {
                   IOB_AXISTREAM_OUT_CSRS_FIFO_LEVEL_W);
 }
 
-uint16_t iob_axistream_out_csrs_get_version() {
+uint32_t iob_axistream_out_csrs_get_version() {
   return iob_read(base + IOB_AXISTREAM_OUT_CSRS_VERSION_ADDR,
                   IOB_AXISTREAM_OUT_CSRS_VERSION_W);
 }
